requirements.txt
fastapi==0.119.1
uvicorn[standard]==0.27.1

# LangChain ecosystem
langchain-core==1.2.20
langchain==1.1.1

langchain-openai==1.1.11

langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8

pydantic>=2.7.4,<3.0.0
pydantic-settings==2.2.1
python-dotenv==1.0.1

sqlalchemy==2.0.29
psycopg2-binary==2.9.9

pandas==2.2.1
matplotlib==3.8.3
scipy==1.17.1

python-jose[cryptography]==3.3.0

passlib[bcrypt]==1.7.4
bcrypt==3.2.2

python-multipart==0.0.9

email-validator==2.1.1

# Docker Python SDK — для запуска sandbox-контейнеров из python_tool.py
docker>=7.0.0

# PostgreSQL checkpointer для LangGraph (персистентная история диалогов).
# psycopg v3 — отдельный пакет, не конфликтует с psycopg2-binary.
# AsyncPostgresSaver использует psycopg3 async API.
psycopg[binary]>=3.1.0
langgraph-checkpoint-postgres>=2.0.0

# Testing
pytest==8.3.2
pytest-asyncio==0.24.0
pytest-mock==3.14.0
httpx==0.27.2
pytest-cov==5.0.0